djshop/requirements/base.txt
Django==4.2.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2
psycopg==3.1.18
channels==4.1.0
channels-redis==4.2.0
Pillow==10.3.0
celery==5.3.6
redis==5.0.3
django-treebeard==4.7.1
django-guardian==2.4.0
